feat: Add store-catalog facade interface DTOs
This commit adds the store-catalog facade interface DTOs, including the input and output DTOs for finding a store catalog, as well as the output DTO for finding all store catalogs.

Refactor the code to include the following changes:
- Add `FindStoreCatalogFacadeInputDto` and `FindStoreCatalogFacadeOutputDto` interfaces
- Add `FindAllStoreCatalogFacadeOutputDto` interface
- Update `StoreCatalogFacadeInterface` to include the new methods

These changes are necessary to support the functionality of the store-catalog module.

Author: Glaucia Lemos
Date: 08/19/2024